chore: moving the get cascading models util to common app
