+++ bc_obps/service/tests/test_report_version_service.py
@@ -1,23 +1,10 @@
-from django.db.models import CASCADE, Field, Model
+from common.tests.utils.model_inspection import get_cascading_models
 from model_bakery import baker
 import pytest
 from reporting.models.report_version import ReportVersion
 from service.report_version_service import ReportVersionService
 
 pytestmark = pytest.mark.django_db
-
-
-def get_cascading_models(model: Model | Field) -> set[Model]:
-    if not hasattr(model, "_meta"):
-        return {}
-
-    delete_cascade_models = {
-        r.related_model for r in model._meta.get_fields() if hasattr(r, "on_delete") and r.on_delete == CASCADE
-    }
-    nested_models_list = [get_cascading_models(m) for m in delete_cascade_models]
-    nested_models = {item for nested_list in nested_models_list for item in nested_list}
-
-    return {*delete_cascade_models, *nested_models}
 
 
 class TestReportVersionService:
